simplify finding the user's Desktop directory
this previous code explicitely parsed the user-dirs.dirs in
~/.config, but would have failed if this location has been
remapped. Furthermore, it relied on manual parsing of this
file.

The current code replaces this with a call to
	GLib.get_user_special_dir(GLib.UserDirectory.DIRECTORY_DESKTOP)
which takes remapping of ~/.config into account.

The code is originally from
https://github.com/smurphos/nemo_actions_and_cinnamon_scripts/blob/beta/patcher/fileUtils_3.8_0.1.patch

We reload the special dirs cache using
	GLib.reload_user_special_dirs_cache()
to get the same behaviour to the previous code which loaded the
file from disk.